// src/app/p-app/p-marketing/shared/dto/DTOMAConfig.dto.ts
export class DTOMAConfig {
    Code: number = 0
    ConfigName: string = ""
    ConfigURL: string = ''
    ConfigDescription: string = ''
    ImageSetting1: string = ''
    ImageSetting2: string = ''
    ImageSetting3: string = ''
    ImageSetting4: string = ''
    ImageSetting5: string = ''
    ExtendConfig: string = ''
    Remark: string = ''
    TypeDataName: string = ''
    TypeData: number = 1
    StatusID: number = 0
    StatusName: string = 'Đang soạn thảo'
    CreateTime: Date | string = null
    CreateBy: string = ''
    LastModifiedTime: Date | string = null
    LastModifiedBy: string = ''
}